pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "circuitron"
version = "0.1.0"
description = "Agentic PCB design accelerator converting natural language prompts into SKiDL code and schematics."
readme = "README.md"
authors = [{name = "Shaurya Sethi"}]
requires-python = ">=3.10"
dependencies = [
    "openai-agents==0.2.10",
    "python-dotenv>=1.1.0",
    "skidl>=2.0.1",
    "httpx>=0.27.0",
    "prompt_toolkit>=3.0",
    "rich>=13.0",
    "logfire>=3.22.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.49.0"]


[project.scripts]
circuitron = "circuitron.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["circuitron*"]

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
addopts = "-q"

